+++ packages/video_player/video_player_platform_interface/test/video_player_platform_interface_test.dart
@@ -40,4 +40,44 @@ void main() {
   test('default implementation isAudioTrackSupportAvailable returns false', () {
     expect(initialInstance.isAudioTrackSupportAvailable(), false);
   });
+
+  test(
+    'default implementation startPictureInPicture throws unimplemented',
+    () async {
+      await expectLater(
+        () => initialInstance.startPictureInPicture(1),
+        throwsUnimplementedError,
+      );
+    },
+  );
+
+  test(
+    'default implementation stopPictureInPicture throws unimplemented',
+    () async {
+      await expectLater(
+        () => initialInstance.stopPictureInPicture(1),
+        throwsUnimplementedError,
+      );
+    },
+  );
+
+  test(
+    'default implementation isPictureInPictureSupported throws unimplemented',
+    () async {
+      await expectLater(
+        () => initialInstance.isPictureInPictureSupported(1),
+        throwsUnimplementedError,
+      );
+    },
+  );
+
+  test(
+    'default implementation isPictureInPictureActive throws unimplemented',
+    () async {
+      await expectLater(
+        () => initialInstance.isPictureInPictureActive(1),
+        throwsUnimplementedError,
+      );
+    },
+  );
 }